requirements.txt
# ------------------------------------------------------------ #
#                        Data Processing                       #
# ------------------------------------------------------------ #

pandas
Pillow
scipy

# ------------------------------------------------------------ #
#                          ML Packages                         #
# ------------------------------------------------------------ #

scikit-learn
opencv-python

# ------------------------------------------------------------ #
#                            PyTorch                           #
# ------------------------------------------------------------ #

-f https://download.pytorch.org/whl/cu117
torch
-f https://download.pytorch.org/whl/cu117
torchtext
-f https://download.pytorch.org/whl/cu117
torchvision

# ------------------------------------------------------------ #
#                           Frontend                           #
# ------------------------------------------------------------ #

gradio
si-prefix